refatoração de código em toda aplicação

+++ src/modules/Produto/useCases/ProdutoUseCase.ts
@@ -1,5 +1,8 @@
 import IRepository from "../../../repositories/IRepository";
 import { IProduto } from "../../../models/Produto";
+import Subcategoria from "../../../models/Subcategoria";
+import BadRequest from "../../../infra/erros/BadRequest";
+import { ObjectId } from "mongoose";
 
 type PayloadCadastroProduto = IProduto;
 
@@ -9,14 +12,38 @@ export default class CadastroUseCase {
     this.repository = cadastroRepository;
   }
   
-  criar(payload: PayloadCadastroProduto){
+  async criar(subcategoriaId: string, payload: PayloadCadastroProduto){
     const produtoData = {
       nome: payload.nome,
       descricao: payload.descricao,
       preco: payload.preco,
       imagem: payload.imagem
     }
-    return this.repository.create(produtoData);
+
+    const subcategoria = await Subcategoria.findById({
+      _id: subcategoriaId,
+    });
+
+    if (!subcategoria) {
+      throw new BadRequest("Subcategoria não encontrada", 400);
+    };
+
+    const produto = await this.repository.create(produtoData);
+
+    let produtoExistente: IProduto[] | ObjectId[] = [];
+
+    if (subcategoria) {
+      produtoExistente = subcategoria.produtos;
+    };
+
+    await Subcategoria.findByIdAndUpdate(subcategoriaId, {
+      produtos: [
+        ...produtoExistente,
+        produto._id
+      ],
+    });
+
+    return produto;
   }
 
   listar() {
@@ -24,14 +51,23 @@ export default class CadastroUseCase {
   }
 
   listarId(id: any) {
-    return this.repository.findOne(id);
+    if (!id) {
+      throw new BadRequest("id inválido!", 400);
+    }
+    return this.repository.findOne({_id: id});
   }
 
   atualizar(id: any, payload: PayloadCadastroProduto) {
+    if (!id) {
+      throw new BadRequest("id inválido!", 400);
+    }
     return this.repository.update(id, payload);
   }
 
   deletar(id: any) {
+    if (!id) {
+      throw new BadRequest("id inválido!", 400);
+    }
     return this.repository.deleteOne(id);
   }
 }
